% Generated by roxygen2 (4.0.1): do not edit by hand
\name{approxeq}
\alias{approxeq}
\title{APPROXEQ Are a and b approximately equal (to within a specified tolerance)?
p = approxeq(a, b, thresh)
'tol' defaults to 1e-3.}
\usage{
approxeq(a, b, tol = 0.001)
}
\description{
APPROXEQ Are a and b approximately equal (to within a specified tolerance)?
p = approxeq(a, b, thresh)
'tol' defaults to 1e-3.
}

