% Generated by roxygen2 (4.0.1): do not edit by hand
\name{assign.cols}
\alias{assign.cols}
\title{Assigns all columns of df into variables with the same name in environment env}
\usage{
assign.cols(df, dest = sys.frame(sys.parent(1)))
}
\description{
Assigns all columns of df into variables with the same name in environment env
}

