% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc.mean.from.F.fun}
\alias{calc.mean.from.F.fun}
\title{Calculate numerically the expected value given a cdf}
\usage{
calc.mean.from.F.fun(F.fun, x.min = 0, x.max = Inf, abs.tol = 10^(-10),
  x.seq = NULL, use.num.integrate = TRUE, ...)
}
\description{
Calculate numerically the expected value given a cdf
}

