% Generated by roxygen2 (4.0.1): do not edit by hand
\name{check.global.vars}
\alias{check.global.vars}
\title{Some functions that are useful for coding
Looks through all loaded functions and searches for
global variables that are used within the functions
this is a common source for errors}
\usage{
check.global.vars()
}
\description{
Some functions that are useful for coding
Looks through all loaded functions and searches for
global variables that are used within the functions
this is a common source for errors
}

