% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clone.environment}
\alias{clone.environment}
\title{Clones an environment and its children}
\usage{
clone.environment(env, made.clones = as.environment(list(org = list(), copy =
  list())), clone.parents = TRUE, clone.global = FALSE, exclude = NULL,
  clone.children = TRUE)
}
\description{
Clones an environment and its children
}

