% Generated by roxygen2 (4.0.1): do not edit by hand
\name{discretize.given.F.vec}
\alias{discretize.given.F.vec}
\title{Helper function to discretize a continous distribution.
F.vec is a finite vector containing the value of the cdf at M different points.
The function generates an M dimension vector of probabilities summing up to 1
that discretize the distribution}
\usage{
discretize.given.F.vec(F.vec)
}
\description{
Helper function to discretize a continous distribution.
F.vec is a finite vector containing the value of the cdf at M different points.
The function generates an M dimension vector of probabilities summing up to 1
that discretize the distribution
}

