% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findzero}
\alias{findzero}
\title{Finds position where the function f becomes zero
First tries find.root and if this fails tries optimize}
\usage{
findzero(f, lower, upper, tol = .Machine$double.eps * 10, result.tol = tol,
  try.uniroot = TRUE, ...)
}
\description{
Finds position where the function f becomes zero
First tries find.root and if this fails tries optimize
}

