% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grid.matrix.permutation}
\alias{grid.matrix.permutation}
\title{Gives the corresponding rows for a permutated grid.matrix given a permutation
x.perm of the elements of the original list x}
\usage{
grid.matrix.permutation(x, perm.col)
}
\description{
Gives the corresponding rows for a permutated grid.matrix given a permutation
x.perm of the elements of the original list x
}

