% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grid.to.matrix}
\alias{grid.to.matrix}
\title{Transforms a grid in long format into a matrix}
\usage{
grid.to.matrix(grid, nrow = length(unique(grid[, 1])),
  ncol = length(unique(grid[, 2])), val.col = 3)
}
\description{
Transforms a grid in long format into a matrix
}

