% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matrix.to.grid}
\alias{matrix.to.grid}
\title{Some functions that are useful for manipulating or creating matrices and data.frames and working with lists of vectors, lists of lists or lists of matrices
Transforms a matrix into grid in long format}
\usage{
matrix.to.grid(mat, x = 1:NROW(mat), y = 1:NCOL(mat), x.name = "x",
  y.name = "y")
}
\description{
Some functions that are useful for manipulating or creating matrices and data.frames and working with lists of vectors, lists of lists or lists of matrices
Transforms a matrix into grid in long format
}

