% Generated by roxygen2 (4.0.1): do not edit by hand
\name{named.list}
\alias{named.list}
\title{Some functions that are useful for lists and environments in particular generating, transforming, copying and assigning values
Creates a list that is named by the names of its arguments}
\usage{
named.list(...)
}
\description{
Some functions that are useful for lists and environments in particular generating, transforming, copying and assigning values
Creates a list that is named by the names of its arguments
}

