% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.multi.lines}
\alias{plot.multi.lines}
\title{Plot several lines}
\usage{
\method{plot}{multi.lines}(mat = NULL, xvar, yvar, ynames = yvar,
  col = NULL, ylim = NULL, xlab = xvar, ylab = "", legend.pos = NULL,
  legend.title = NULL, add = FALSE, lwd = 1, ...)
}
\description{
Plot several lines
}

