% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rbind.list}
\alias{rbind.list}
\title{rbinds a list of matrices, a list of lists, or a list of vectors into a data.frame (or matrix)
each column is a list
Assume that all columns in the sublists are in the same order}
\usage{
rbind.list(li, cols = NULL, check.common.cols = FALSE)
}
\description{
rbinds a list of matrices, a list of lists, or a list of vectors into a data.frame (or matrix)
each column is a list
Assume that all columns in the sublists are in the same order
}

