% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sk.levelplot}
\alias{sk.levelplot}
\title{My wrapper to the lattice function levelplot. Allows for some own color schemes
The parameter focus specifies at which z range stronger color changes shall appear}
\usage{
sk.levelplot(x = NULL, y = NULL, z = NULL, xnames = NULL,
  ynames = NULL, grid.xyz = NULL, col.scheme = "darkredgreen",
  na.col = NULL, at = NULL, at.scheme = "interval", focus = 0,
  cuts = 15, col.regions = NULL, xlab = NULL, ylab = NULL,
  panel = panel.levelplot, zlim = NULL, reverse.colors = FALSE, ...)
}
\description{
My wrapper to the lattice function levelplot. Allows for some own color schemes
The parameter focus specifies at which z range stronger color changes shall appear
}

