% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sk.optim}
\alias{sk.optim}
\title{A wrapper for optimization. Allows to specify which variables shall be free
Has the same syntax for one and multidimensional optmization
Uses optim, omptimize or a grid search}
\usage{
sk.optim(par, f, lower = NULL, upper = NULL, free.par = 1:NROW(par),
  method = "default", num.grid.steps = NULL, maximize = TRUE,
  f.can.take.matrix = FALSE, tol = .Machine$double.eps^0.25, ...)
}
\description{
A wrapper for optimization. Allows to specify which variables shall be free
Has the same syntax for one and multidimensional optmization
Uses optim, omptimize or a grid search
}

