% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sk.pareto.frontier}
\alias{sk.pareto.frontier}
\title{Calculates the 2dimensional paretofrontier of the points val1 and val2
The function returns the indices of the points that lie on the Pareto Frontier
ordered by val1 and val2.}
\usage{
sk.pareto.frontier(val1, val2, tol = 0, ord = NULL)
}
\description{
Calculates the 2dimensional paretofrontier of the points val1 and val2
The function returns the indices of the points that lie on the Pareto Frontier
ordered by val1 and val2.
}

