% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.colMaxs}
\alias{which.colMaxs}
\title{Computes quickly the index of the largest element of each column of a matrix}
\usage{
which.colMaxs(mat)
}
\description{
Computes quickly the index of the largest element of each column of a matrix
}

