% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.colMins}
\alias{which.colMins}
\title{Computes quickly the index of the smallest element of each column of a matrix}
\usage{
which.colMins(mat)
}
\description{
Computes quickly the index of the smallest element of each column of a matrix
}

