% Generated by roxygen2 (4.0.1): do not edit by hand
\name{which.rowMins}
\alias{which.rowMins}
\title{Computes quickly the index of the smallest element of each row of a matrix}
\usage{
which.rowMins(mat)
}
\description{
Computes quickly the index of the smallest element of each row of a matrix
}

